Handle 404s from upstream HBase properly
Properly pass down 404s received from HBase instead of just handling any
non-200 response from HBase as an application error and responding to the
client with an HTTP 500.

Change the port to listen on to 8001 due to conflict with appsync daemon that
made it impossible to run both on the same machine for testing.

Removing trailing whitespace because it's the fruit of the devil.